Split mod_p4_d.rs (grew past 800 lines from the pattern-1 router decompositions) to clear max-lines
