+++ crates/almide-mir/src/lower/mod_c.rs
@@ -662,6 +662,7 @@ include!("mod_p4.rs");
 include!("mod_p4_b.rs");
 include!("mod_p4_c.rs");
 include!("mod_p4_d.rs");
+include!("mod_p4_e.rs");
 include!("mod_p5.rs");
 include!("mod_p5_b.rs");
 // The desugar family (formerly one 4.8k-line mod_p6.rs), split by concern:
